--- backend/techpick-batch/src/main/java/kernel360/techpick/batch/domain/rss/service/RssService.java
@@ -1,13 +1,0 @@
-package kernel360.techpick.batch.domain.rss.service;
-
-import java.util.List;
-
-import kernel360.techpick.batch.domain.rss.dto.RssBlogCommand;
-import kernel360.techpick.batch.domain.rss.dto.RssBlogResult;
-
-public interface RssService {
-
-	RssBlogResult saveRssBlog(RssBlogCommand.Create command);
-
-	List<RssBlogResult> getAllRssBlog();
-}
