+++ backend/techpick-batch/src/main/java/techpick/TechPickBatchApplication.java
@@ -1,10 +1,10 @@
-package kernel360;
+package techpick;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication(
-	scanBasePackages = {"kernel360.techpick.batch", "kernel360.techpick.core"}
+	scanBasePackages = {"techpick.batch", "techpick.core"}
 )
 public class TechPickBatchApplication {
 	public static void main(String[] args) {
