+++ backend/techpick-batch/src/main/java/techpick/batch/domain/rss/service/RssService.java
@@ -1,0 +1,13 @@
+package techpick.batch.domain.rss.service;
+
+import java.util.List;
+
+import techpick.batch.domain.rss.dto.RssBlogCommand;
+import techpick.batch.domain.rss.dto.RssBlogResult;
+
+public interface RssService {
+
+	RssBlogResult saveRssBlog(RssBlogCommand.Create command);
+
+	List<RssBlogResult> getAllRssBlog();
+}
